requirements.txt
torch>=2.2
numpy>=1.24,<2.0
metadrive-simulator>=0.4.2
gymnasium==0.29.*
supabase>=2.0
python-dotenv>=1.0
mediapy>=1.2
matplotlib>=3.7
tensorboard>=2.14
pytest>=7.0
